% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{accessors}
\alias{accessors}
\alias{model}
\alias{model,segmentation-method}
\alias{emission}
\alias{emission,segmentation-method}
\alias{transition}
\alias{transition,segmentation-method}
\alias{overlap}
\alias{overlap,segmentation-method}
\alias{TSS}
\alias{TSS,segmentation-method}
\alias{TES}
\alias{TES,segmentation-method}
\alias{segment}
\alias{segment,segmentation-method}
\alias{bins}
\alias{bins,segmentation-method}
\alias{counts}
\alias{counts,segmentation-method}
\alias{likelihood}
\alias{likelihood,segmentation-method}
\alias{cells}
\alias{cells,segmentation-method}
\alias{states}
\alias{states,segmentation-method}
\alias{markers}
\alias{markers,segmentation-method}
\title{Accessors for the \code{segmentation} objects}
\usage{
model(object)

\S4method{model}{segmentation}(object)

emission(object)

\S4method{emission}{segmentation}(object)

transition(object)

\S4method{transition}{segmentation}(object)

overlap(object, ...)

\S4method{overlap}{segmentation}(object, cell)

TSS(object, ...)

\S4method{TSS}{segmentation}(object, cell)

TES(object, ...)

\S4method{TES}{segmentation}(object, cell)

segment(object, ...)

\S4method{segment}{segmentation}(object, cell)

bins(object, ...)

\S4method{bins}{segmentation}(object, cell)

counts(object, ...)

\S4method{counts}{segmentation}(object, cell)

likelihood(object)

\S4method{likelihood}{segmentation}(object)

cells(object)

\S4method{cells}{segmentation}(object)

states(object)

\S4method{states}{segmentation}(object)

markers(object)

\S4method{markers}{segmentation}(object)
}
\arguments{
\item{object}{An object of class \code{segmentation}}

\item{...}{Other argument passed to the accessors}

\item{cell}{A string}
}
\description{
These functions can be used to access the contents of \code{segmentation}
objects as well as modifying them.
}
\seealso{
segmentation
}
