% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenteR.R
\name{ad_hoc_reorder}
\alias{ad_hoc_reorder}
\title{Reorder the sections inside section_title_df}
\usage{
ad_hoc_reorder(section_title_df)
}
\arguments{
\item{section_title_df}{A dataframe with the sections title that has been
found inside the article, their position inside the output of poppler, the
font and the size of the font.
Word                         Font                 Size
293     Introduction VMUQDX+ITCStoneSans-Semibold 10.0
1321         Results VMUQDX+ITCStoneSans-Semibold 10.0}
}
\value{

}
\description{
If the position of the section "Acknowledgments" is set before "Conclusion",
reorder the sections titles to position "Acknowledgments" after "Conclusion".
This would otherwise lead to crash downstream since Conclusion is placed
before for a human reader AND for tabulizer, which handle(s) double column
writing style better than poppler.
}
\details{
In "Smulders, S et al 2015.pdf" the problem was encountered :section_title_dfWord                   Font   Size543      Introduction CLNCIJ+AdvOT18499c10.B 7.97021089        Materials CLNCIJ+AdvOT18499c10.B 7.97021091          Methods CLNCIJ+AdvOT18499c10.B 7.97021787          Results CLNCIJ+AdvOT18499c10.B 7.97022263       Discussion CLNCIJ+AdvOT18499c10.B 7.97023447 Acknowledgements CLNCIJ+AdvOT18499c10.B 7.97023449       Conclusion CLNCIJ+AdvOT18499c10.B 7.97023624    Supplementary CLNCIJ+AdvOT18499c10.B 7.97023642       References CLNCIJ+AdvOT18499c10.B 7.9702
}
