% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_sections_position_in_ConllU.R
\name{adding_missing_first_letter_section}
\alias{adding_missing_first_letter_section}
\title{Send back occurrences of a truncated name of the section title}
\usage{
adding_missing_first_letter_section(conllu_df, section, occurrences)
}
\arguments{
\item{conllu_df}{A dataframe of the raw text of the article annotated using
the model, in the CoNLL-U format.}

\item{section}{A string, a section name.}

\item{occurrences}{A int, an index inside the conllu_df dataframe. This index
refer to a row, which correspond to a token, and all the information related
to it such as : lemma, sentence, head_token, etc.}
}
\value{
Occurrences
}
\description{
Please also refer to the function missing_first_letter_section().
adding_missing_first_letter_section() is called before
missing_first_letter_section() in handle_typos().
#' In one of article of Elsevier, "Al-Bairuty, G et al 2013.pdf",
"Acknowledgements" became "cknowledgements" and "Reference", "eference".
This function search for tokens that correspond to truncated name of
the section title. If there is occurrences for truncated versions, the function
send back to the main functions the new occurrences.
}
