% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenteR.R
\name{annotate_txt_pdf}
\alias{annotate_txt_pdf}
\title{Download the model, Tokenize and make the Dependency Parsing of the extracted text}
\usage{
annotate_txt_pdf(txt_pdf, udpipe_model)
}
\arguments{
\item{txt_pdf}{Raw text read from the article in pdf, extracted using the
tabulapdf::extract_text() function.}

\item{udpipe_model}{An object of class udpipe_model, as returned by udpipe_load_model()
from the package udpipe. If left void, the function will download the model
english-gum annotation model from the treebanks available at
https://universaldependencies.org and load it to annotate the raw text.}
}
\value{
A dataframe of the raw text of the article annotated using the model, in the CoNLL-U format.
Please refer to https://bnosac.github.io/udpipe/en/index.html to know more about it.
}
\description{
Download the model, Tokenize and make the Dependency Parsing of the extracted text
}
