% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{binarize_bed}
\alias{binarize_bed}
\title{Binarize the bed files}
\usage{
binarize_bed(
  inputdir,
  cellmarkfiletable,
  chromsizefile,
  binsize = 200,
  outputdir
)
}
\arguments{
\item{inputdir}{A string. The dirctory of the bam files.}

\item{cellmarkfiletable}{A string. The path to the input files table. Only}

\item{chromsizefile}{A string. The path to the chromosomes sizes file.}

\item{binsize}{An integer. The number in bp used to generate binarized files.}

\item{outputdir}{A string. The path to a directory where output will be
written.}
}
\value{
NULL. Write files to the outputdir
}
\description{
Transform the aligned reads into a binary format.
}
\seealso{
Binarize binarize_bam
}
