% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_typos.R
\name{capitalize_first_letter}
\alias{capitalize_first_letter}
\title{Capitalize the first letter and lower the other letters}
\usage{
capitalize_first_letter(section)
}
\arguments{
\item{section}{A string, a section name.}
}
\value{
A string with the first letter capitalized and the other one lowered.
}
\description{
Used to correct strange behavior of tabulizer for section title.
For example, "MAterIAls"->"Materials".
Must be used on must on conllu_df$token.
}
