% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenteR.R
\name{check_sections_df}
\alias{check_sections_df}
\title{Message if a section appeared in duplicate in positions_sections_df}
\usage{
check_sections_df(positions_sections_df)
}
\arguments{
\item{positions_sections_df}{A dataframe with the positions of the sections
currently localized inside the conllu_df.}
}
\description{
Send a message if there is duplicate in positions_sections_df.
}
