% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenteR.R
\name{clean_font_txt}
\alias{clean_font_txt}
\title{Remove the most abundant font, usually the font of the text}
\usage{
clean_font_txt(poppler_output)
}
\arguments{
\item{poppler_output}{A dataframe with the text box (font), name of the font
and the size of the font, produced by prepare_poppler_output.}
}
\value{

}
\description{
In "Baker, G L et al 2008.pdf", the word of the section's titles had a smaller
size than the text of the article. This seems to happened in old articles/for
different fonts that have a different scale.
Removing the font of the text is necessary in this case, and, in general way,
calling this function once help processing the text.
This function is called in the main function of the package and
find_section_tile().
}
