% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenteR.R
\name{clean_section_title}
\alias{clean_section_title}
\title{Preprocess the section titles}
\usage{
clean_section_title(positions_sections_df)
}
\arguments{
\item{positions_sections_df}{A dataframe with the positions of the sections
currently localized inside the conllu_df.}
}
\value{

}
\description{
This function preprocess the section title to help the extractions functions.
For example, "aterial" cannot be find in "MATERIAL".
Called in extract_section_from_conllu().
}
