% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenteR.R
\name{clean_title_journal}
\alias{clean_title_journal}
\title{Clean title of the journal from the section_title_df.}
\usage{
clean_title_journal(pdf_name, section_title_df)
}
\arguments{
\item{pdf_name}{A string, the name of the article in pdf format, for which we
want to extract a specific section specified by section_aliases.}

\item{section_title_df}{A dataframe with the sections title that has been
found inside the article, their position inside the output of poppler, the
font and the size of the font.
Word                         Font                 Size
293     Introduction VMUQDX+ITCStoneSans-Semibold 10.0
1321         Results VMUQDX+ITCStoneSans-Semibold 10.0}
}
\value{

}
\description{
An article, Guo, J et al 2014.pdf highlight a the following problem :
the title of the journal can be written with the same font that the sections
titles. The simplest solution to handle this problem is to correct the
section_title_df().
This function check if there is a section repeated in the dataframe a number
of time equal to the number of pages of the pdf or equal to the number of
page less 1.
}
