% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{compare_models}
\alias{compare_models}
\title{Compare two or more models}
\usage{
compare_models(objs, type = "emission", plot = FALSE, ...)
}
\arguments{
\item{objs}{A list of segmentation items}

\item{type}{A string. What to compare. Default to 'emission'}

\item{plot}{A logical.}

\item{...}{Other arguments passed to plot}
}
\value{
A numeric vector or a plot with the same values.
}
\description{
Compare two or more models
}
