% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_reads_ranges}
\alias{count_reads_ranges}
\title{Count reads in \code{GRanges} objects from bam files}
\usage{
count_reads_ranges(ranges, cellmarkfiletable, inputbamdir)
}
\arguments{
\item{ranges}{A \code{GRanges} to count in.}

\item{cellmarkfiletable}{A string. The path to the input files table.}

\item{inputbamdir}{A \code{string}. The path to the input bam files
directory.}
}
\value{
A \code{SummarizedExperiment} object with \code{ranges} as its
\code{rowRanges} and the counts as the \code{assay}.
}
\description{
Count reads in \code{GRanges} objects from bam files
}
