% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_section_titles_dataframe.R
\name{create_section_title_df}
\alias{create_section_title_df}
\title{Find what are the sections and in which order they appear}
\usage{
create_section_title_df(font_section, list_of_sections, poppler_output)
}
\arguments{
\item{list_of_sections}{A list of vector of string, that contain the sections
titles we want to localize in the article, if they exist.}

\item{poppler_output}{A dataframe with the text box (font), name of the font
and the size of the font, produced by prepare_poppler_output.}
}
\value{
A dataframe with the sections title that has been found inside the article,
their position inside the output of poppler, the font and the size of the font.
Word                         Font                 Size
293     Introduction VMUQDX+ITCStoneSans-Semibold 10.0
1321         Results VMUQDX+ITCStoneSans-Semibold 10.0
}
\description{
The function call successively find_section_titles() on each vector of the list.
The function then call re_identify_font_section(), in case Acknowledgment and
References (if they exist) have a different font from the other sections titles.
This function will change the output only if the section_title_df contain 2
or less section.
}
