% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_calls.R
\name{.Binarize}
\alias{.Binarize}
\title{Call Java \code{BinarizeBed}}
\usage{
.Binarize(inputdir, cellmarkfiletable, chromsizefile, binsize, outputdir, type)
}
\arguments{
\item{inputdir}{A string. The path to bed files.}

\item{cellmarkfiletable}{A tab delimited files of three columns.
The columns contains the cell, mark and the name or the bed file.}

\item{chromsizefile}{A string. The path to the chromosomes sizes file.}

\item{binsize}{An integer. The bin size to use. Defualut is 200.}

\item{outputdir}{A string. The path to a directory where output will be
written.}

\item{type}{A string. The file type 'bam' or 'bed'.}
}
\value{
\code{NULL}. Output files are written to the output directory.
}
\description{
Call the Java module \code{BinarizeBed} which binarize a bed file of the
aligned reads.
}
\seealso{
binarize_bed
}
