% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_calls.R
\name{.LearnModel}
\alias{.LearnModel}
\title{Call Java \code{LearnModel}}
\usage{
.LearnModel(
  inputdir,
  outputdir,
  numstates,
  coordsdir,
  anchorsdir,
  chromsizefile,
  assembly,
  optional
)
}
\arguments{
\item{inputdir}{A string. The path to binarized files.}

\item{outputdir}{A string. The path to a directory where output will be
written.}

\item{numstates}{An integer. The number of desired states in the model.}

\item{coordsdir}{A string. The path to genomic coordiantes files.}

\item{anchorsdir}{A string. The path to the genomic anchors files.}

\item{chromsizefile}{A string. The path to the chromosomes sizes file.}

\item{assembly}{A string. The name of the genomic assembely.}

\item{optional}{A string. Other optional arguments passed to the Java
command.}
}
\value{
\code{NULL}. Output files are written to the output directory.
}
\description{
Call the Java module \code{LearnModel} which learns a multi-state model
from ChIP-seq data.
}
\seealso{
learn_model
}
