% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_sections_position_in_ConllU.R
\name{eliminate_cf_occurrences}
\alias{eliminate_cf_occurrences}
\title{Remove the occurrences of section title that are preceded by a cf occurrence}
\usage{
eliminate_cf_occurrences(conllu_df, occurrences)
}
\arguments{
\item{conllu_df}{A dataframe of the raw text of the article annotated using
the model, in the CoNLL-U format.}

\item{occurrences}{A int, an index inside the conllu_df dataframe. This index
refer to a row, which correspond to a token, and all the information related
to it such as : lemma, sentence, head_token, etc.}
}
\value{
Note created during the documentation : this function could be improved :
\itemize{
\item by extending the range to look fot the tokens, to handle sentence such as
"please see the material and method section".
\item including word such as "refer".
}
}
\description{
This function remove the occurrences of a section title that are tokens
preceded by tokens that indicate a confer context (i.e. "compare" or "consult")
such as "see", "cf", "discussed", "described".
Simply speaking, it allow the calling function, locate_sections_position_in_conllu(),
to ignore sentence such as "For more details, please see material and method
section".
Note created during the documentation : this function could be improved :
\itemize{
\item by extending the range to look for the tokens, to handle sentence such as
"please see the material and method section".
\item including word such as "refer".
}
}
