% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{emissions_file}
\alias{emissions_file}
\title{Make emissions file name}
\usage{
emissions_file(numstates)
}
\arguments{
\item{numstates}{An integer}
}
\value{
A string
}
\description{
Make emissions file name
}
