% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{enrichment_files}
\alias{enrichment_files}
\title{Make enrichment file names}
\usage{
enrichment_files(numstates, cells, table = "RefSeq", annotation = "TSS")
}
\arguments{
\item{numstates}{An integer}

\item{cells}{A character vector}

\item{table}{A string}

\item{annotation}{A string}
}
\value{
A character vector
}
\description{
Make enrichment file names
}
