% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{extract_bins}
\alias{extract_bins}
\title{Extract binariezed data for a GRange}
\usage{
extract_bins(bins, which, tidy = FALSE)
}
\arguments{
\item{bins}{A SummarizedExperiment object}

\item{which}{A GRanges object}

\item{tidy}{A logical}
}
\value{
A GRanges object
}
\description{
Extract binariezed data for a GRange
}
