% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenteR.R
\name{extract_section_from_conllu}
\alias{extract_section_from_conllu}
\title{Extract the section from the ConllU dataframe}
\usage{
extract_section_from_conllu(conllu_df, positions_sections_df, section_aliases)
}
\arguments{
\item{conllu_df}{A dataframe of the raw text of the article annotated using
the model, in the CoNLL-U format.}

\item{positions_sections_df}{A dataframe with the positions of the sections
currently localized inside the conllu_df.}

\item{section_aliases}{A vector of strings, which are the different names that
the section extracted by extract_section_from_pdf() can have.}
}
\value{
The section extracted in a dataframe in the CoNLL-U format.
}
\description{
Extract the section specified by the vector of section aliases.
The function use a for loop to parse the different sections in the
position_section_df to find a match. Earlier build stopped at the first match
and returned everything between the start of the section and the next one
inside position_section_df. Now it continues to look for match after the
first one, in case of, for example, "Material" and "Methods" are split into
different sections.
}
