% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenteR.R
\name{extract_section_from_pdf}
\alias{extract_section_from_pdf}
\title{Extract a specified section from a scientific article in pdf format.}
\usage{
extract_section_from_pdf(
  pdf_name,
  udpipe_model,
  section_aliases,
  remove_bibliography = TRUE
)
}
\arguments{
\item{pdf_name}{A string, the name of the article in pdf format, for which we
want to extract a specific section specified by section_aliases.}

\item{udpipe_model}{An object of class udpipe_model, as returned by udpipe_load_model()
from the package udpipe. If left void, the function will download the model
english-gum annotation model from the treebanks available at
https://universaldependencies.org and load it to annotate the raw text.}

\item{section_aliases}{A vector of strings, which are the different names that
the section extracted by extract_section_from_pdf() can have.}

\item{remove_bibliography}{Boolean. If the references of the article must be
removed during the extraction, to avoid error due to aliases of the section
names inside the references. Default to TRUE (references of the articles are removed).}
}
\value{
A dataframe of the section of the article in the CoNLL-U format.
Please refer to https://bnosac.github.io/udpipe/en/index.html to know more about this specific format.
}
\description{
The main function of the package. Taking the name of a pdf, it extract the
section that have a section title that match the section_aliases provided by
the user. Please refer to the vignette of the package for a more detailed
description of the internal working of this function.
}
