% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_section_titles_dataframe.R
\name{find_section_titles}
\alias{find_section_titles}
\title{Find if a section (vector of aliases) exist and where it is located}
\usage{
find_section_titles(vector_title, font_section, poppler_output)
}
\arguments{
\item{vector_title}{A vector of string, that identify the different names of
a section title we aim to localise in the article, such as
c("Conclusions", "Conclusion", "CONCLUSION", "CONCLUSIONS").}

\item{font_section}{The font of the section as identified by identify_font().}

\item{poppler_output}{A dataframe with the text box (font), name of the font
and the size of the font, produced by prepare_poppler_output.}
}
\value{
Position of the title section if exist inside poppler_output.
}
\description{
The function check first if the word exist, and if the word with the maximum
size inside the text has the same font than the font of the titles. If this
fail, clean_font_txt() to remove the smallest font size from poppler_output.
Then recursive call.
}
