% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_distance}
\alias{get_distance}
\title{Get the distances of the segments to the TSS in each cell type}
\usage{
get_distance(segments, average = FALSE)
}
\arguments{
\item{segments}{A \code{GRanges} object. Usually the output of calling
\code{segment} on the the output object of \code{lean_model}.}

\item{average}{A logical. Whether the width should be averaged across cells.}
}
\value{
A \code{data.frame}
}
\description{
Get the distances of the segments to the TSS in each cell type
}
