% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_frequency}
\alias{get_frequency}
\title{Get the frequency of the segments in each cell type}
\usage{
get_frequency(segments, normalize = FALSE, tidy = FALSE, plot = FALSE, ...)
}
\arguments{
\item{segments}{A \code{GRanges} object. Usually the output of calling
\code{segment} on the the output object of \code{lean_model}.}

\item{normalize}{A logical. Whether the frequency should be normalized by
the total number of segments}

\item{tidy}{A logical.}

\item{plot}{A logical.}

\item{...}{Other arguments passed to barplot}
}
\value{
A \code{data.frame} when tidy is TRUE otherwise a matrix or a plot
}
\description{
Get the frequency of the segments in each cell type
}
