% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenteR.R
\name{grep_head_token}
\alias{grep_head_token}
\title{Grep the lemma corresponding to the head token of the token at index}
\usage{
grep_head_token(conllu_df, index)
}
\arguments{
\item{conllu_df}{A dataframe of the raw text of the article annotated using
the model, in the CoNLL-U format.}

\item{index}{}
}
\value{

}
\description{
Grep the lemma corresponding to the head_token_id of the token at the entry
at index in the conllu_df. Function called by eliminate_cf_occurrences().
}
