% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_typos.R
\name{handle_typos}
\alias{handle_typos}
\title{Solve problems due to typo inside the ConLL-U dataframe}
\usage{
handle_typos(conllu_df, section, occurrences)
}
\arguments{
\item{conllu_df}{A dataframe of the raw text of the article annotated using
the model, in the CoNLL-U format.}

\item{section}{A string, a section name.}

\item{occurrences}{A int, an index inside the conllu_df dataframe. This index
refer to a row, which correspond to a token, and all the information related
to it such as : lemma, sentence, head_token, etc.}
}
\value{

}
\description{
This function call various functions to solve a list of problems encountered
in locate_sections_position_in_conllu() due to typos generated by the extraction of the
text directly from the pdf. NB : this functions only work on the NLP dataframe,
in ConLL-U format. This function call :
\itemize{
\item missing_first_letter_section() (when "Acknowledgements" became
"cknowledgements", and "Reference", "eference")
\item capitalize_first_letter() ("MAterIAls"->"Materials", correct strange
behavior of tabulizer for section title)
\item regex_correction() (when "Acknowledgements" became "group.Acknowledgments")
}
}
