% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenteR.R
\name{identify_font}
\alias{identify_font}
\title{Identify the font used for the section titles}
\usage{
identify_font(poppler_output)
}
\arguments{
\item{poppler_output}{A dataframe with the text box (font), name of the font
and the size of the font, produced by prepare_poppler_output.}
}
\value{
The name of the font used for the sections titles.
}
\description{
This function identify the most likely font used to write the sections title
inside the article in pdf, by looking first at the font used in Acknowledgment
and References.
The function first try to identify where are the words References and
Acknowledgements in the poppler output. After that function try to identify if
there is a section named References, and it not, Acknowledgement
}
