% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_sections_position_in_ConllU.R
\name{is_summary_box}
\alias{is_summary_box}
\title{Detect and handle the summary box}
\usage{
is_summary_box(conllu_df, section, occurrences, section_title_df)
}
\arguments{
\item{conllu_df}{A dataframe of the raw text of the article annotated using
the model, in the CoNLL-U format.}

\item{section}{A string, a section name.}

\item{occurrences}{A int, an index inside the conllu_df dataframe. This index
refer to a row, which correspond to a token, and all the information related
to it such as : lemma, sentence, head_token, etc.}

\item{section_title_df}{A dataframe with the sections title that has been
found inside the article, their position inside the output of poppler, the
font and the size of the font.
Word                         Font                 Size
293     Introduction VMUQDX+ITCStoneSans-Semibold 10.0
1321         Results VMUQDX+ITCStoneSans-Semibold 10.0}
}
\value{

}
\description{
In "Berce, C et al 2016.pdf", but also in many other article, there is short
summary of the article in a box at the beginning of the article with sections
names. Depending of the content of this "summary box", this can lead the main
function of the package to fail to get meaningful result or to extract anything
at all. The function focus on the "introduction" section of the summary box,
since the function will ignore the box if the section introduction is absent
from the summary box : the function will localize the Introduction section,
usually after the summary box, and then ignore the summary box when looking
for the other section due to the subset_occurrence() function.
}
