% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_sections_position_in_ConllU.R
\name{locate_sections_position_in_conllu}
\alias{locate_sections_position_in_conllu}
\title{Localize the position of the sections inside the Conll-U dataframe}
\usage{
locate_sections_position_in_conllu(conllu_df, section_title_df)
}
\arguments{
\item{conllu_df}{A dataframe of the raw text of the article annotated using
the model, in the CoNLL-U format.}

\item{section_title_df}{A dataframe with the sections title that has been
found inside the article, their position inside the output of poppler, the
font and the size of the font.
Word                         Font                 Size
293     Introduction VMUQDX+ITCStoneSans-Semibold 10.0
1321         Results VMUQDX+ITCStoneSans-Semibold 10.0}
}
\value{
The position of the section in the Conllu data structure.
}
\description{
This function create and return the a dataframe with the name of the section
and its starting position inside conllu_df. Simple structure of the function :
}
\details{
locate_sections_position_in_conllu() For section name in list of section (identify by the fonts information)
lower_but_first_letter(token of conllu_df in section)
lower_but_first_letter(token of conllu_df in lower_but_first_letter(section))
eliminate_cf_occurrences()
adding_missing_first_letter_section()
multiple_section_fix()
subset_occurrences()
handle_typos()
is_summary_box()
}
