% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_sections_position_in_ConllU.R
\name{lower_but_first_letter}
\alias{lower_but_first_letter}
\title{Lower all the letter except the first one}
\usage{
lower_but_first_letter(token)
}
\arguments{
\item{token}{A string of character, which is a token inside the ConLL-U
dataframe.}
}
\value{
A string of character corrected.
}
\description{
Correct strange behavior of tabulizer for section titles, by lowering all the
letters except the first one which is left uncorrected.
"MAterIAls" become "Materials". This function allow to find tokens such as
"MAterIAls" when searching for "Material", while avoiding to find token such
as "material".
}
