% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenteR.R
\name{merging_section}
\alias{merging_section}
\title{Merge sections name into one, such as "Discussion and Results"}
\usage{
merging_section(positions_sections_df)
}
\arguments{
\item{positions_sections_df}{A dataframe with the positions of the sections
currently localized inside the conllu_df.}
}
\value{
The position of the section in the Conllu data structure.
}
\description{
This function aims to merge really close section in positions_section_df
The problem arise from section title like "results and discussion", or
"materials and methods".
The way the results of the poppler is read does not discriminate for section
in the same sentence. A section "results and discussion" would create two
entries in positions_sections_df. This function gather together sections in
positions_sections_df that occurred at a distance of two words or so.
}
