% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_typos.R
\name{missing_first_letter_section}
\alias{missing_first_letter_section}
\title{Look for token that are a substring of section.}
\usage{
missing_first_letter_section(conllu_df, section)
}
\arguments{
\item{conllu_df}{A dataframe of the raw text of the article annotated using
the model, in the CoNLL-U format.}

\item{section}{A string, a section name.}
}
\value{
occurences of truncated section such as "eference"
}
\description{
This function is called in handle_typos(), after adding_missing_first_letter_section().
In one of article of Elsevier, "Al-Bairuty, G et al 2013.pdf",
"Acknowledgements" became "cknowledgements", and "Reference", "eference"
but this only occurred in the conllu_df, Conll-U datastructure from UDpipe.
}
