% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{model_file}
\alias{model_file}
\title{Make model file name}
\usage{
model_file(numstates)
}
\arguments{
\item{numstates}{An integer}
}
\value{
A string
}
\description{
Make model file name
}
\examples{
model_file(3)

}
