% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{overlap_files}
\alias{overlap_files}
\title{Make overlap file names}
\usage{
overlap_files(numstates, cells)
}
\arguments{
\item{numstates}{An integer}

\item{cells}{A character vector}
}
\value{
A character vector
}
\description{
Make overlap file names
}
