% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_font_infos.R
\name{prepare_poppler_output}
\alias{prepare_poppler_output}
\title{Extract the font information from an article in pdf}
\usage{
prepare_poppler_output(pdf_name)
}
\arguments{
\item{pdf_name}{A string, the name of the article in pdf format, for which we
want to extract a specific section specified by section_aliases.}
}
\value{
A dataframe with the text box (font), name of the font and the size of the font.
}
\description{
Extract the font information from an article in pdf
}
