% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenteR.R
\name{preprocess_article_txt}
\alias{preprocess_article_txt}
\title{Preprocess the text of the article using regex expression}
\usage{
preprocess_article_txt(txt_pdf)
}
\arguments{
\item{txt_pdf}{The text of the article, into a single string, extracted from
the pdf using tabulapdf}
}
\value{

}
\description{
This function is a succession of regex to correct/preprocess the text extracted
from the pdf. Since the raw text is extracted directly from the pdf using
tabulapdf, there is some typos or recurrent character so remove. For example,
non graphical caracter, or "-\" and newline character in the middle of a word,
or concatenated at the beginning of the word.
}
