% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{range_bins}
\alias{range_bins}
\title{Format the loaded binarized data}
\usage{
range_bins(bins, chromsizefile, binsize, return = "GRanges", tidy = TRUE)
}
\arguments{
\item{bins}{A \code{list} of the \code{read_bins_file} output.}

\item{chromsizefile}{A string. The path to the chromosomes sizes file.}

\item{binsize}{An integer. The number in bp used to generate binarized files.}

\item{return}{A string. Possible values are \code{GRanges} (default) or
\code{SummarizedExperiment}.}

\item{tidy}{A \code{logical}. Default is \code{TRUE}. Whether to tidy the
metadata columns of the \code{GRanges} object.}
}
\value{
\code{GRanges} (default) or \code{SummarizedExperiment}.
}
\description{
The function takes the \code{data.frame}s of the loaded binarized data files
and format them into \code{GRanges} or \code{SummarizedExperiment} objects.
}
