% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_section_titles_dataframe.R
\name{re_identify_font_section}
\alias{re_identify_font_section}
\title{Re identify the font for the sections titles}
\usage{
re_identify_font_section(poppler_output, section_title_df, list_of_sections)
}
\arguments{
\item{poppler_output}{A dataframe with the text box (font), name of the font
and the size of the font, produced by prepare_poppler_output.}

\item{section_title_df}{A dataframe with the sections title that has been
found inside the article, their position inside the output of poppler, the
font and the size of the font.
Word                         Font                 Size
293     Introduction VMUQDX+ITCStoneSans-Semibold 10.0
1321         Results VMUQDX+ITCStoneSans-Semibold 10.0}

\item{list_of_sections}{A list of vector of string, that contain the sections
titles we want to localize in the article, if they exist.}
}
\value{

}
\description{
When Abstract and References has a different font than the other section
(it happens in some journals). This function check if the section_df is too
small (length =< 2) and looks for the font of "Introduction", "Abstract" or
"Background" in poppler output (if it exist in one occurrence only).
The function next call find_section_titles() with the new font, and return
the new section_title_df merge with the existing one to the calling function,
create_section_title_df().
Function initially made for "Campagnolo, L et al 2013.pdf".
}
