% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_bins_file}
\alias{read_bins_file}
\title{Read \code{bins} files}
\usage{
read_bins_file(file)
}
\arguments{
\item{file}{A string. The path to the file.}
}
\value{
A \code{list} of 3 items: cell, seqname and binaries.
}
\description{
The files contain the cell and the chromosome info in the first line and
the binarized data from all marks in the rest.
}
\examples{
# locate the file
fl <- system.file('extdata/SAMPLEDATA_HG18/',
                  'GM12878_chr11_binary.txt.gz',
                  package = 'segmenter')

# read the file
read_bins_file(fl)

}
