% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_cellmark_file}
\alias{read_cellmark_file}
\title{Read \code{cellmarktable} file}
\usage{
read_cellmark_file(file)
}
\arguments{
\item{file}{A string. The path to the file.}
}
\value{
A \code{data.frame}
}
\description{
The file should contain at least three columns: cell, mark and file for the
names of the cells/conditions, the available marks and binarized data files.
}
\examples{
# locate the file
fl <- system.file('extdata/input',
                  'cell_mark_table.tsv',
                  package = 'segmenter')

# read the file
read_cellmark_file(fl)

}
