% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_chromsize_file}
\alias{read_chromsize_file}
\title{Read \code{chromsizefile}}
\usage{
read_chromsize_file(file)
}
\arguments{
\item{file}{A string. The path to the file.}
}
\value{
A \code{data.frame}
}
\description{
The file should contain exactly two columns. One for the name of the
chromosome and the other for its length.
}
\examples{
# locate the file
chromsizefile <- system.file('extdata/ChromHMM/CHROMSIZES',
                             'hg18.txt',
                             package = 'segmentr')

# read the file
read_chromsize_file(chromsizefile)

}
