% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_emissions_file}
\alias{read_emissions_file}
\title{Read \code{emissions} file}
\usage{
read_emissions_file(file, states, marks)
}
\arguments{
\item{file}{A string. The path to the file.}

\item{states}{A \code{character} vector. The names of the states.}

\item{marks}{A \code{character} vector. The names of the marks}
}
\value{
A \code{matrix}
}
\description{
The segments files are the output of running \code{learn_model} and named
\code{emissions_3_segment.bed}
}
\examples{
# locate the file
fl <- system.file('extdata/output',
                           'emissions_3.txt',
                           package = 'segmenter')

# read the file
read_emissions_file(fl)

}
