% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_enrichment_file}
\alias{read_enrichment_file}
\title{Read \code{enrichment} files}
\usage{
read_enrichment_file(file, states, regions)
}
\arguments{
\item{file}{A string. The path to the file.}

\item{states}{A \code{character} vector. The names of the states.}

\item{regions}{A \code{character} vector. The names of the regions.}
}
\value{
A \code{matrix}
}
\description{
The segments files are the output of running \code{learn_model} and named
\code{<cell>_3_TSS.txt} or \code{<cell>_3_TES.txt}.
}
\examples{
# locate the file
fl <- system.file('extdata/output',
                  'GM12878_3_RefSeqTSS_neighborhood.txt',
                  package = 'segmenter')

# read the file
read_enrichment_file(fl)

}
