% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_model_file}
\alias{read_model_file}
\title{Read \code{modelfile}}
\usage{
read_model_file(file)
}
\arguments{
\item{file}{A string. The path to the file.}
}
\value{
A \code{data.frame}
}
\description{
The model file is the output of running \code{learn_model} and named
\code{model_#.txt}
}
\examples{
# locate the file
modelfile <- file.path(tempdir(), 'model_3.txt')

# read the file
read_model_file(modelfile)

}
