% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_segements_file}
\alias{read_segements_file}
\title{Read \code{segments} files}
\usage{
read_segements_file(file, states)
}
\arguments{
\item{file}{A string. The path to the file.}

\item{states}{A \code{character} vector. The names of the states.}
}
\value{
A \code{data.frame}
}
\description{
The segments files are the output of running \code{learn_model} and named
\code{<cell>_3_segment.bed}
}
\examples{
# locate the file
segmentfile <- system.file('extdata/output',
                           'GM12878_3_segments.bed',
                           package = 'segmenter')

# read the file
segs <- read_segements_file(segmentfile)
head(segs)

}
