% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_transitions_file}
\alias{read_transitions_file}
\title{Read \code{transitions} file}
\usage{
read_transitions_file(file, states)
}
\arguments{
\item{file}{A string. The path to the file.}

\item{states}{A \code{character} vector. The names of the states.}
}
\value{
A \code{matrix}
}
\description{
The segments files are the output of running \code{learn_model} and named
\code{transitions_3_segment.bed}
}
\examples{
# locate the file
fl <- system.file('extdata/output',
                           'transitions_3.txt',
                           package = 'segmentr')

# read the file
read_transitions_file(fl)

}
