% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_sections_position_in_ConllU.R
\name{recursive_filter_first_lemma}
\alias{recursive_filter_first_lemma}
\title{Check if the first lemma of the sentence is the lemma at the position index}
\usage{
recursive_filter_first_lemma(conllu_df, index, lemma_nb)
}
\arguments{
\item{conllu_df}{A dataframe of the raw text of the article annotated using
the model, in the CoNLL-U format.}

\item{index}{A int. Similar to occurrences, indicate the position of a token
and all the information related to it such as : lemma, sentence, head_token,
etc. Use here to not create confusion with the variable occurrence.}

\item{lemma_nb}{Variable used for the recursive call, allow to advance from
one lemma to the next in the sentence between each call.}
}
\value{

}
\description{
Called by select_first_lemma_sentence(). Check if the first lemma of the
sentence is the lemma at the position index. The function is a recursive call,
the recursion lasting as long as the regex expressions are true.
The regex allow to handle section's title such as "1) Material and methods",
since "1" and ")" would be annotated as token/lemma by udpipe, the goal of the
function being to identify if the index corresponding to "Material" is the
first lemma of the sentence.
When the recursive call is made, the lemma at the index is compared to the
following lemma of the sentence, compared to the previous call.
The regex are : detect punctuation, detect decimal, detect character string
"Animals|Animal" (to handle a really particular article).
}
