% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_sections_position_in_ConllU.R
\name{reduce_occurrences}
\alias{reduce_occurrences}
\title{Reduce the number of potential localization of section titles}
\usage{
reduce_occurrences(
  conllu_df,
  occurrences,
  positions_sections_df,
  section_title_df
)
}
\arguments{
\item{conllu_df}{A dataframe of the raw text of the article annotated using
the model, in the CoNLL-U format.}

\item{occurrences}{A int, an index inside the conllu_df dataframe. This index
refer to a row, which correspond to a token, and all the information related
to it such as : lemma, sentence, head_token, etc.}

\item{positions_sections_df}{A dataframe with the positions of the sections
currently localized inside the conllu_df.}

\item{section_title_df}{A dataframe with the sections title that has been
found inside the article, their position inside the output of poppler, the
font and the size of the font.
Word                         Font                 Size
293     Introduction VMUQDX+ITCStoneSans-Semibold 10.0
1321         Results VMUQDX+ITCStoneSans-Semibold 10.0}
}
\value{

}
\description{
This function reduce the number of occurrences that can be a localization of
section titles. This function successively call subset_occurrences(),
select_first_token_sentence_if_section_title() and select_first_lemma_sentence().
}
