% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenteR.R
\name{remove_bibliography_from_conllu}
\alias{remove_bibliography_from_conllu}
\title{Remove the bibliography from conllu.}
\usage{
remove_bibliography_from_conllu(conllu_df, section_title_df)
}
\arguments{
\item{conllu_df}{A dataframe of the raw text of the article annotated using
the model, in the CoNLL-U format.}

\item{section_title_df}{A dataframe with the sections title that has been
found inside the article, their position inside the output of poppler, the
font and the size of the font.
Word                         Font                 Size
293     Introduction VMUQDX+ITCStoneSans-Semibold 10.0
1321         Results VMUQDX+ITCStoneSans-Semibold 10.0}
}
\value{
the conllu_df without the whole bibliography
}
\description{
This function must be called before remove_reference_section_from_titles.
This function remove the bibliography from the article.
This is to prevent failure due to have sections aliases located in the
references section, such as material in a title of an article.
This can cause the whole script to fail, since no other section would be found
after this falsely located section.
}
