% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenteR.R
\name{remove_reference_section_from_titles}
\alias{remove_reference_section_from_titles}
\title{Remove the references section}
\usage{
remove_reference_section_from_titles(section_title_df)
}
\arguments{
\item{section_title_df}{A dataframe with the sections title that has been
found inside the article, their position inside the output of poppler, the
font and the size of the font.
Word                         Font                 Size
293     Introduction VMUQDX+ITCStoneSans-Semibold 10.0
1321         Results VMUQDX+ITCStoneSans-Semibold 10.0}
}
\value{
section_title_df without the reference and the following section
}
\description{
This function remove the "references" section and the ones that came after in
section_title_df. This is to prevent failure due to have sections aliases
located in the references section. This can cause the whole script to fail,
since no other section would be found after this falsely located section.
}
