% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{segments_files}
\alias{segments_files}
\title{Make segments file names}
\usage{
segments_files(numstates, cells)
}
\arguments{
\item{numstates}{An integer}

\item{cells}{A character vector}
}
\value{
A character vector
}
\description{
Make segments file names
}
\examples{
segments_files(3, 'K562')

}
