% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_sections_position_in_ConllU.R
\name{select_first_lemma_sentence}
\alias{select_first_lemma_sentence}
\title{Return an occurrence is first lemma of a sentence}
\usage{
select_first_lemma_sentence(conllu_df, occurrences)
}
\arguments{
\item{conllu_df}{A dataframe of the raw text of the article annotated using
the model, in the CoNLL-U format.}

\item{occurrences}{A int, an index inside the conllu_df dataframe. This index
refer to a row, which correspond to a token, and all the information related
to it such as : lemma, sentence, head_token, etc.}
}
\value{
A lesser list of occurrences
}
\description{
This function return the first occurrence passed as input which is the first
lemma of a sentence. Typically, it distinguish between sentence "northern blot
as describe in Materials and Methods" and sentence like "Materials and Methods",
the second one being the materials and methods section.
}
