% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_sections_position_in_ConllU.R
\name{select_first_token_sentence_if_section_title}
\alias{select_first_token_sentence_if_section_title}
\title{Return first token of the sentence if inside section title df}
\usage{
select_first_token_sentence_if_section_title(
  conllu_df,
  occurrences,
  positions_sections_df
)
}
\arguments{
\item{conllu_df}{A dataframe of the raw text of the article annotated using
the model, in the CoNLL-U format.}

\item{occurrences}{A int, an index inside the conllu_df dataframe. This index
refer to a row, which correspond to a token, and all the information related
to it such as : lemma, sentence, head_token, etc.}

\item{positions_sections_df}{A dataframe with the positions of the sections
currently localized inside the conllu_df.}
}
\value{

}
\description{
This function call select_if_firsts_tokens_already_in_section_title() for
each of the index passed as input.
It returns the first occurrence for which the first token of the sentence is
already in the dataframe section_title_df.
}
\details{
For example, if there is a Material and Methods section, and several time the
words methods in the article, the others functions cannot distinguished which
occurrence to select (since it is not the first lemma of a sentence).
}
