% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_sections_position_in_ConllU.R
\name{select_if_firsts_tokens_already_in_section_title}
\alias{select_if_firsts_tokens_already_in_section_title}
\title{Send true if the first token of the sentence is already inside the section titles}
\usage{
select_if_firsts_tokens_already_in_section_title(
  conllu_df,
  index,
  positions_sections_df
)
}
\arguments{
\item{conllu_df}{A dataframe of the raw text of the article annotated using
the model, in the CoNLL-U format.}

\item{index}{A int, position of a token inside the conllu_df.}

\item{positions_sections_df}{A dataframe with the positions of the sections
currently localized inside the conllu_df.}
}
\value{

}
\description{
This function check if the first token is already in positions_sections_df.
}
