% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_sections_position_in_ConllU.R
\name{subset_occurrences}
\alias{subset_occurrences}
\title{Reduce the search space of section titles}
\usage{
subset_occurrences(occurrences, positions_sections_df)
}
\arguments{
\item{occurrences}{A int, an index inside the conllu_df dataframe. This index
refer to a row, which correspond to a token, and all the information related
to it such as : lemma, sentence, head_token, etc.}

\item{positions_sections_df}{A dataframe with the positions of the sections
currently localized inside the conllu_df.}
}
\value{

}
\description{
Simply speaking, it reduce the search of section names to portion of the
article after the already annotated section titles. In Abrams et al 2010,
"Introduction" is at position 338, "Results" at 1501, "Discussion" at 5975 and
there is occurrences of "Materials" at 1799, 2747, 3816, 7055.
The three firsts ones are likely part of the "Results" section.
}
