% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{test_obj}
\alias{test_obj}
\title{A segmentation object generated from the test data}
\format{
An object of class \code{segmentation} of length 1.
}
\usage{
test_obj
}
\description{
A segmentation object generated by running \code{lean_model} on the test
dataset in 'inst/extdata/ChromHMM/SAMPLEDATA_HG18'. The source code to
this run is in 'inst/script/test_obj.R'
}
\keyword{datasets}
