% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{test_objs}
\alias{test_objs}
\title{A a list of segmentation objects generated from the test data}
\format{
An object of class \code{list} of length 6.
}
\usage{
test_objs
}
\description{
A segmentation object generated by running \code{lean_model} on the test
dataset in inst/extdata/ChromHMM/SAMPLEDATA_HG18' for 3 to 8 states.
The source code to this run is in 'data-raw/test_obj.R'
}
\keyword{datasets}
