% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_ranges}
\alias{tidy_ranges}
\title{Tidy the metadata of a \code{GRanges} object}
\usage{
tidy_ranges(gr, columns, low = 0)
}
\arguments{
\item{gr}{A \code{GRanges} object}

\item{columns}{A \code{character} vectors. The names of columns to be tidied.}

\item{low}{An \code{integer}. All values <= this \code{integer} will be
removed.}
}
\value{
A \code{GRanges} object
}
\description{
Tidy the metadata of a \code{GRanges} object
}
