% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{transitions_file}
\alias{transitions_file}
\title{Make transitions file name}
\usage{
transitions_file(numstates)
}
\arguments{
\item{numstates}{An integer}
}
\value{
A string
}
\description{
Make transitions file name
}
