% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{methods}
\alias{methods}
\alias{show,segmentation-method}
\title{Methods to interact with \code{segmentation} objects}
\usage{
\S4method{show}{segmentation}(object)
}
\arguments{
\item{object}{An object of class \code{segmentation}}
}
\description{
These functions can be used to interact with \code{segmentation} objects for
purposes other than accessing or modifying their contents.
}
\seealso{
segmentation

accessors
}
